ec_signature splitted in separate files with more focused content: utils, ecdsa, ecssa ecsignto contract
